Fix type of default key when creating BPF maps
The default key (0) used to populate various BPF maps was declared as
uint32_t. but the 'dvars' BPF map uses a uint64_t key.  Since we know
that the value of the key will always be 0, it is safe to just make it
uint64_t for all maps, and have it be interpreted as uint32_t for the
maps that use a smaller key size.

+++ libdtrace/dt_bpf.c
@@ -227,7 +227,7 @@ dt_bpf_gmap_create(dtrace_hdl_t *dtp)
 	int		stabsz, gvarsz, lvarsz, aggsz, memsz;
 	int		dvarc = 0;
 	int		ci_mapfd, st_mapfd, pr_mapfd;
-	uint32_t	key = 0;
+	uint64_t	key = 0;
 	size_t		strsize = dtp->dt_options[DTRACEOPT_STRSIZE];
 	uint8_t		*buf, *end;
 	char		*strtab;
